Move ClocksourceManager.h out of ConfigurationManager.h to ChronoVisorServer2.h and client.h
