+++ ChronoAPI/ChronoLog/include/ClocksourceManager.h
@@ -11,25 +11,11 @@
 #include <typeinfo>
 #include <unistd.h>
 #include <emmintrin.h>
+#include <enum.h>
 #include <log.h>
 
 #define   lfence()  _mm_lfence()
 #define   mfence()  _mm_mfence()
-
-enum ClocksourceType {
-    C_STYLE = 0,
-    CPP_STYLE = 1,
-    TSC = 2
-};
-
-inline const char* getClocksourceTypeString(ClocksourceType type) {
-    switch (type) {
-        case C_STYLE: return "C_STYLE";
-        case CPP_STYLE: return "CPP_STYLE";
-        case TSC: return "TSC";
-        default: return "UNKNOWN";
-    }
-}
 
 class Clocksource {
 public:
